Fixed Room CRUDS

+++ Code/src/main/java/com/hms/hms_dashboard_01/controller/tab/RoomController.java
@@ -1,5 +1,6 @@
 package com.hms.hms_dashboard_01.controller.tab;
 
+import com.hms.hms_dashboard_01.DTO.RoomDTO;
 import com.hms.hms_dashboard_01.dal.DALRoomManager;
 import com.hms.hms_dashboard_01.model.entities.Room;
 import com.hms.hms_dashboard_01.utility.path;
@@ -11,6 +12,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
@@ -42,43 +44,52 @@ public class RoomController implements Initializable  {
 
 
 //    Holds the data for the table view in the room tab of the admin/warden
-    ObservableList<Room> roomList = FXCollections.observableArrayList(
-            new Room(1, "Available", 1000, "None", "1st Floor"),
-            new Room(2, "Available", 1000, "Ahmed", "1st Floor"),
-            new Room(3, "Available", 1000, "St2", "1st Floor"),
-            new Room(4, "Unavailable", 1000, "St8", "1st Floor")
-    );
+    ObservableList<Room> roomList = FXCollections.observableArrayList(DALRoomManager.getAllRooms());
 // Assign the values to the columns of the table view in the room tab of the admin/warden
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
+        loadTable();
+    }
+    public void loadTable(){
         roomNo.setCellValueFactory(new PropertyValueFactory<>("roomNo"));
 //        roomType.setCellValueFactory(new PropertyValueFactory<>("roomType"));
         roomFee.setCellValueFactory(new PropertyValueFactory<>("roomFee"));
         roomStatus.setCellValueFactory(new PropertyValueFactory<>("roomStatus"));
         roomAllocatedTo.setCellValueFactory(new PropertyValueFactory<>("roomAssignedTo"));
-        roomLocation.setCellValueFactory(new PropertyValueFactory<>("roomLocation"));
+//        roomLocation.setCellValueFactory(new PropertyValueFactory<>("roomLocation"));
 
         roomTable.setItems(roomList);
-
-
-
-
     }
-
 
     public void addRoom(ActionEvent e) throws IOException {
 //       set the stage scene to add room
         Stage stage = new Stage();
-        Parent root1 = FXMLLoader.load(getClass().getResource(path.getPath("tab", "add_room_form")));
+        FXMLLoader loader = new FXMLLoader(getClass().getResource(path.getPath("tab", "add_room_form")));
+        Parent root1 = loader.load();
+        RoomFormController roomFormController = loader.getController();
+        roomFormController.setRoomController(this);
         stage.setTitle("Add Room");
         stage.setScene(new Scene(root1, 1054, 650));
         stage.show();
+    }
+    public void updateRoomTable() {
+        roomList.clear();
+        roomList.addAll(DALRoomManager.getAllRooms());
     }
 
     public void deleteRoom(ActionEvent e) {
 //        Get selected object from the table view
         Room room = roomTable.getSelectionModel().getSelectedItem();
+        if(room == null){
+            Alert alert = new Alert(Alert.AlertType.ERROR);
+            alert.setTitle("Error");
+            alert.setHeaderText("No room selected");
+            alert.setContentText("Please select a room to delete");
+            alert.showAndWait();
+            return;
+        }
         DALRoomManager.DeleteRoom(room.getRoomNo());
+        updateRoomTable();
     }
 
 }
